refactor(match): moves handleDuplicatedSupportRequest to separate func
