--- src/lib/emailDuplicated.ts
@@ -1,14 +1,0 @@
-export default function emailDuplicated(firstName: string) {
-	return `Olá, ${firstName}!
- 
-Nossa equipe identificou que você solicitou ajuda em nosso site mais de uma vez. No momento do seu primeiro cadastro, nós enviamos uma resposta por email sobre o seu atendimento. Por isso, queremos saber, como podemos te ajudar?
- 
-Se você não recebeu o encaminhamento necessário e ainda precisa de apoio, não se preocupe! Por favor, nos escreva relatando o ocorrido para que a nossa equipe saiba como te auxiliar para que seja acolhida. 
- 
-Aguardamos o seu contato,
-
-Estamos juntas! 
- 
-Um abraço,
-Equipe do Mapa do Acolhimento`;
-}
